+++ .pylintrc
@@ -7,11 +7,11 @@ extension-pkg-whitelist= PyQt6
 
 # Add files or directories to the blacklist. They should be base names, not
 # paths.
-ignore=CVS,mock,options_ui.py,weather_ui.py
+ignore=CVS,mock
 
 # Add files or directories matching the regex patterns to the blacklist. The
 # regex matches against base names, not paths.
-ignore-patterns=options.ui.py,weather_ui.py
+ignore-patterns=*_ui.py
 
 # Python code to execute, usually for sys.path manipulation such as
 # pygtk.require().
