+++ src/conan_app_launcher/ui/views/app_grid/config/__init__.py
@@ -11,7 +11,7 @@ from conan_app_launcher import INVALID_CONAN_REF, PathLike
 UI_CONFIG_JSON_TYPE = "json"
 
 if TYPE_CHECKING:
-    from ..views.app_grid.model import UiAppLinkModel, UiTabModel
+    from ..model import UiAppLinkModel, UiTabModel
 
 
 @dataclass
